+++ src/main/java/com/iexec/core/config/OpenApiConfig.java
@@ -16,19 +16,20 @@
 
 package com.iexec.core.config;
 
-import com.iexec.core.version.VersionService;
 import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.info.Info;
+import org.springframework.boot.info.BuildProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
 public class OpenApiConfig {
 
-    private final VersionService versionService;
+    public static final String TITLE = "iExec Core Scheduler";
+    private final BuildProperties buildProperties;
 
-    public OpenApiConfig(VersionService versionService) {
-        this.versionService = versionService;
+    public OpenApiConfig(BuildProperties buildProperties) {
+        this.buildProperties = buildProperties;
     }
 
     /*
@@ -38,8 +39,8 @@ public class OpenApiConfig {
     public OpenAPI api() {
         return new OpenAPI().info(
                 new Info()
-                        .title("iExec Core Scheduler")
-                        .version(versionService.getVersion())
+                        .title(TITLE)
+                        .version(buildProperties.getVersion())
         );
     }
 }